Markdown formatters

Format a character vector with Markdown.

Usage

md_bold(string)

md_italic(string)

md_strikethrough(string)

md_underline(string)

md_spoiler(string)

Arguments

string

A character vector.

Value

A character vector with the same length as the input.

Details

The Markdown formatters in the gluedown package use glue, which support strings and not character vectors. For consistency, since all string-altering functions in this package support and return character vectors, vectorised Markdown-formatters are also included. The Markdown syntax used here is tailored for Discord.